added verification
-Added verification if statement, that if the Admin HAS completed the
textbox response, and has updated it to the server via PUT
method(supportResponseData), it should place that value within the
textbox, and change the button id attribute to something else making it
unable to send off new responses, as well as display "Disabled" over the
buttons text.
-Also added the opposite verification. If there is NO user support
request data added to the id of the Admin Response page, then the page
will display a console.log & error notification(via Toastr), that states
this SupportRequestData's ID does not exist. Not sure if this will be
necessary if we link the supprt requests page directly to the admin
response page. Preventing anyone from just placing a supportRequestData
ID and needing this verification of the existence of said ID.