Passed in a function w/in the filter parameter
Passed in a function w/in the filter parameter of my angularJS function.
Previous response.item.email doesn't display email due to it having a
null reference.
Attempts:
1.) Tried creating a seperate function within the controller and passing
this within the function directly 'main.verifyEmail', but this was a no
go.
2.) Tried passing function w/in an ng=show, ng-hide, as well as an ng-if
directive, but these just check for truthy and falsey values, or the
execution of my function then returns the expression IF these conditions
are met. SO,...FAIL.